[#1049] Added dialog to download form
